requirements.txt
#Packages needed
streamlit
pandas
plotly
mysql-connector-python